Try avoiding emitting assembly to run setup process
