+++ Mono.Addins/Mono.Addins.Database/SetupProcess.cs
@@ -68,10 +68,10 @@ namespace Mono.Addins.Database
 			sb.Append (" \"").Append (arg1).Append ("\"");
 
 			Process process = new Process ();
-			
-			string asm = null;
+
+			string thisAsm = typeof (SetupProcess).Assembly.Location;
+			string asm = Path.GetDirectoryName (thisAsm) + "Mono.Addins.SetupProcess.dll";
 			try {
-				asm = CreateHostExe ();
 				if (!Util.IsMono)
 					process.StartInfo = new ProcessStartInfo (asm, sb.ToString ());
 				else {
@@ -107,79 +107,7 @@ namespace Mono.Addins.Database
 			} catch (Exception ex) {
 				Console.WriteLine (ex);
 				throw;
-			} finally {
-				if (asm != null) {
-					try {
-						File.Delete (asm);
-					} catch { }
-				}
 			}
-		}
-
-		public static int Main (string [] args)
-		{
-			ProcessProgressStatus monitor = new ProcessProgressStatus (int.Parse (args [0]));
-
-			try {
-				string registryPath = Console.In.ReadLine ();
-				string startupDir = Console.In.ReadLine ();
-				string addinsDir = Console.In.ReadLine ();
-				string databaseDir = Console.In.ReadLine ();
-
-				AddinDatabase.RunningSetupProcess = true;
-				AddinRegistry reg = new AddinRegistry (registryPath, startupDir, addinsDir, databaseDir);
-
-				switch (args [1]) {
-				case "scan": {
-						string folder = args.Length > 2 ? args [2] : null;
-						if (folder.Length == 0) folder = null;
-
-						var context = new ScanOptions ();
-						context.Read (Console.In);
-						reg.ScanFolders (monitor, folder, context);
-						break;
-					}
-				case "pre-scan": {
-						string folder = args.Length > 2 ? args [2] : null;
-						if (folder.Length == 0) folder = null;
-						var recursive = bool.Parse (Console.In.ReadLine ());
-						reg.GenerateScanDataFilesInProcess (monitor, folder, recursive);
-						break;
-					}
-				case "get-desc":
-					var outFile = Console.In.ReadLine ();
-					reg.ParseAddin (monitor, args [2], args [3]);
-					break;
-				}
-			} catch (Exception ex) {
-				monitor.ReportError ("Unexpected error in setup process", ex);
-				return 1;
-			}
-			return 0;
-		}
-		
-		static string CreateHostExe ()
-		{
-			AssemblyName aname = new AssemblyName ();
-			aname.Name = Guid.NewGuid ().ToString ().Replace ('-','_');
-			AssemblyBuilder ab = AssemblyBuilder.DefineDynamicAssembly (aname, AssemblyBuilderAccess.Run);
-			ModuleBuilder mb = ab.DefineDynamicModule (aname.Name);
-			TypeBuilder tb = mb.DefineType ("App", TypeAttributes.Public|TypeAttributes.Class);
-			
-			MethodBuilder fb = tb.DefineMethod("Main",
-			                                   MethodAttributes.Public |
-			                                   MethodAttributes.Static,
-			                                   typeof(int), new Type[] { typeof(string[]) });
-			
-			MethodInfo mi = typeof(SetupProcess).GetMethod ("Main", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static);
-			
-			ILGenerator ilg = fb.GetILGenerator();
-			ilg.Emit (OpCodes.Ldarg_0);
-			ilg.EmitCall (OpCodes.Call, mi, null);
-			ilg.Emit (OpCodes.Ret);
-			
-			tb.CreateTypeInfo();
-			return tb.FullName;
 		}
 	}
